--- src/skeleton/mod.rs
@@ -1,22 +1,0 @@
-mod imp;
-
-use glib::Object;
-use gtk::glib;
-
-glib::wrapper! {
-    pub struct Skeleton(ObjectSubclass<imp::Skeleton>)
-        @extends adw::Bin, gtk::Widget,
-        @implements gtk::Accessible, gtk::Buildable, gtk::ConstraintTarget;
-}
-
-impl Skeleton {
-    pub fn new() -> Self {
-        Object::builder().build()
-    }
-}
-
-impl Default for Skeleton {
-    fn default() -> Self {
-        Self::new()
-    }
-}
