+++ src/widgets/mod.rs
@@ -1,0 +1,8 @@
+mod skeleton;
+pub use self::skeleton::Skeleton;
+
+mod basic_numpad;
+pub use self::basic_numpad::BasicNumpad;
+
+mod window;
+pub use self::window::Window;
